+++ test/test.js
@@ -8,17 +8,31 @@ const fs = require( 'fs' );
 
 process.chdir( __dirname );
 
+const getChunksFromGenerated = generated => {
+	if (generated.output) {
+		return generated.output.length ? generated.output : Object.keys(generated.output)
+			.map(chunkName => generated.output[chunkName]);
+	} else {
+		return [generated];
+	}
+};
+
 function executeBundle ( bundle ) {
 	return bundle.generate({
 		format: 'cjs'
 	}).then( generated => {
-		const fn = new Function ( 'module', 'exports', 'assert', generated.code );
+		const fn = new Function ( 'module', 'exports', 'assert', getChunksFromGenerated(generated)[0].code );
 		const module = { exports: {} };
 
 		fn( module, module.exports, assert );
 
 		return module;
 	});
+}
+
+function getBundleImports ( bundle) {
+	return bundle.imports ? Promise.resolve(bundle.imports) : bundle.generate({format: 'esm'})
+		.then(generated => getChunksFromGenerated(generated)[0].imports);
 }
 
 describe( 'rollup-plugin-node-resolve', function () {
@@ -102,7 +116,7 @@ describe( 'rollup-plugin-node-resolve', function () {
 				format: 'cjs'
 			});
 		}).then( generated => {
-			assert.ok( ~generated.code.indexOf( 'setPrototypeOf' ) );
+			assert.ok( ~getChunksFromGenerated(generated)[0].code.indexOf( 'setPrototypeOf' ) );
 		});
 	});
 
@@ -160,7 +174,7 @@ describe( 'rollup-plugin-node-resolve', function () {
 
 	it( 'allows use of object browser field, resolving `main`', function () {
 		return rollup.rollup({
-			entry: 'samples/browser-object-main/main.js',
+			input: 'samples/browser-object-main/main.js',
 			plugins: [
 				nodeResolve({
 					mainFields: [ 'browser', 'main' ]
@@ -175,7 +189,7 @@ describe( 'rollup-plugin-node-resolve', function () {
 
 	it( 'DEPRECATED: options.browser = true still works', function () {
 		return rollup.rollup({
-			entry: 'samples/browser-object-main/main.js',
+			input: 'samples/browser-object-main/main.js',
 			plugins: [
 				nodeResolve({
 					browser: true
@@ -190,7 +204,7 @@ describe( 'rollup-plugin-node-resolve', function () {
 
 	it( 'allows use of object browser field, resolving implicit `main`', function () {
 		return rollup.rollup({
-			entry: 'samples/browser-object/main-implicit.js',
+			input: 'samples/browser-object/main-implicit.js',
 			plugins: [
 				nodeResolve({
 					mainFields: [ 'browser', 'main' ]
@@ -203,7 +217,7 @@ describe( 'rollup-plugin-node-resolve', function () {
 
 	it( 'allows use of object browser field, resolving replaced builtins', function () {
 		return rollup.rollup({
-			entry: 'samples/browser-object-builtin/main.js',
+			input: 'samples/browser-object-builtin/main.js',
 			plugins: [
 				nodeResolve({
 					mainFields: [ 'browser', 'main' ]
@@ -216,7 +230,7 @@ describe( 'rollup-plugin-node-resolve', function () {
 
 	it( 'allows use of object browser field, resolving nested directories', function () {
 		return rollup.rollup({
-			entry: 'samples/browser-object-nested/main.js',
+			input: 'samples/browser-object-nested/main.js',
 			plugins: [
 				nodeResolve({
 					mainFields: [ 'browser', 'main' ]
@@ -231,7 +245,7 @@ describe( 'rollup-plugin-node-resolve', function () {
 
 	it( 'allows use of object browser field, resolving `main`', function () {
 		return rollup.rollup({
-			entry: 'samples/browser-object-main/main.js',
+			input: 'samples/browser-object-main/main.js',
 			plugins: [
 				nodeResolve({
 					mainFields: [ 'browser', 'main' ]
@@ -246,7 +260,7 @@ describe( 'rollup-plugin-node-resolve', function () {
 
 	it( 'allows use of object browser field, resolving implicit `main`', function () {
 		return rollup.rollup({
-			entry: 'samples/browser-object/main-implicit.js',
+			input: 'samples/browser-object/main-implicit.js',
 			plugins: [
 				nodeResolve({
 					mainFields: [ 'browser', 'main' ]
@@ -259,7 +273,7 @@ describe( 'rollup-plugin-node-resolve', function () {
 
 	it( 'allows use of object browser field, resolving replaced builtins', function () {
 		return rollup.rollup({
-			entry: 'samples/browser-object-builtin/main.js',
+			input: 'samples/browser-object-builtin/main.js',
 			plugins: [
 				nodeResolve({
 					mainFields: [ 'browser', 'main' ]
@@ -272,7 +286,7 @@ describe( 'rollup-plugin-node-resolve', function () {
 
 	it( 'allows use of object browser field, resolving nested directories', function () {
 		return rollup.rollup({
-			entry: 'samples/browser-object-nested/main.js',
+			input: 'samples/browser-object-nested/main.js',
 			plugins: [
 				nodeResolve({
 					mainFields: [ 'browser', 'main' ]
@@ -287,7 +301,7 @@ describe( 'rollup-plugin-node-resolve', function () {
 
 	it( 'allows use of object browser field, resolving `main`', function () {
 		return rollup.rollup({
-			entry: 'samples/browser-object-main/main.js',
+			input: 'samples/browser-object-main/main.js',
 			plugins: [
 				nodeResolve({
 					mainFields: [ 'browser', 'main' ]
@@ -302,7 +316,7 @@ describe( 'rollup-plugin-node-resolve', function () {
 
 	it( 'allows use of object browser field, resolving implicit `main`', function () {
 		return rollup.rollup({
-			entry: 'samples/browser-object/main-implicit.js',
+			input: 'samples/browser-object/main-implicit.js',
 			plugins: [
 				nodeResolve({
 					mainFields: [ 'browser', 'main' ]
@@ -315,7 +329,7 @@ describe( 'rollup-plugin-node-resolve', function () {
 
 	it( 'allows use of object browser field, resolving replaced builtins', function () {
 		return rollup.rollup({
-			entry: 'samples/browser-object-builtin/main.js',
+			input: 'samples/browser-object-builtin/main.js',
 			plugins: [
 				nodeResolve({
 					mainFields: [ 'browser', 'main' ]
@@ -328,7 +342,7 @@ describe( 'rollup-plugin-node-resolve', function () {
 
 	it( 'allows use of object browser field, resolving nested directories', function () {
 		return rollup.rollup({
-			entry: 'samples/browser-object-nested/main.js',
+			input: 'samples/browser-object-nested/main.js',
 			plugins: [
 				nodeResolve({
 					mainFields: [ 'browser', 'main' ]
@@ -360,9 +374,8 @@ describe( 'rollup-plugin-node-resolve', function () {
 					preferBuiltins: true
 				})
 			]
-		}).then( bundle => {
-			assert.deepEqual( bundle.imports.sort(), [ 'events' ] );
-		});
+		}).then(getBundleImports)
+			.then(imports => assert.deepEqual(imports, ['events']));
 	});
 
 	it( 'preferBuiltins: false allows resolving a local module with the same name as a builtin module', () => {
@@ -373,24 +386,20 @@ describe( 'rollup-plugin-node-resolve', function () {
 					preferBuiltins: false
 				})
 			]
-		}).then( bundle => {
-			assert.deepEqual( bundle.imports.sort(), [] );
-		});
+		}).then(getBundleImports)
+			.then(imports => assert.deepEqual(imports, []));
 	});
 
 	it( 'issues a warning when preferring a builtin module without having explicit configuration', () => {
 		let warning = null;
 		return rollup.rollup({
 			input: 'samples/prefer-builtin/main.js',
-			plugins: [
-				nodeResolve({
-					onwarn ( message ) {
-						if ( ~message.indexOf( 'prefer' ) ) {
-							warning = message;
-						}
-					}
-				})
-			]
+			onwarn ({message}) {
+				if ( ~message.indexOf( 'preferring' ) ) {
+					warning = message;
+				}
+			},
+			plugins: [nodeResolve()]
 		}).then( () => {
 			const localPath = path.join(__dirname, 'node_modules/events/index.js');
 			assert.strictEqual(
@@ -555,16 +564,16 @@ describe( 'rollup-plugin-node-resolve', function () {
 	});
 
 	it( 'throws error if local id is not resolved', () => {
-		const entry = path.join( 'samples', 'unresolved-local', 'main.js' );
+		const input = path.join( 'samples', 'unresolved-local', 'main.js' );
 		return rollup.rollup({
-			entry,
+			input,
 			plugins: [
 				nodeResolve()
 			]
 		}).then( () => {
 			throw Error( 'test should fail' );
 		}, err => {
-			assert.equal( err.message, `Could not resolve './foo' from ${entry}` );
+			assert.equal( err.message, `Could not resolve './foo' from ${input}` );
 		});
 	});
 
@@ -574,9 +583,8 @@ describe( 'rollup-plugin-node-resolve', function () {
 			plugins: [ nodeResolve({
 				jail: `${__dirname}/samples/`
 			}) ]
-		}).then( (bundle) => {
-			assert.deepEqual(bundle.imports, [ 'string/uppercase.js' ]);
-		});
+		}).then(getBundleImports)
+			.then(imports => assert.deepEqual(imports, ['string/uppercase.js']));
 	});
 
 	it( 'bundle module defined inside the jail', () => {
@@ -585,9 +593,8 @@ describe( 'rollup-plugin-node-resolve', function () {
 			plugins: [ nodeResolve({
 				jail: `${__dirname}/`
 			}) ]
-		}).then( (bundle) => {
-			assert.deepEqual(bundle.imports, []);
-		});
+		}).then(getBundleImports)
+			.then(imports => assert.deepEqual(imports, []));
 	});
 
 	it( '"only" option allows to specify the only packages to resolve', () => {
@@ -598,9 +605,8 @@ describe( 'rollup-plugin-node-resolve', function () {
 					only: [ 'test' ]
 				})
 			]
-		}).then(bundle => {
-			assert.deepEqual( bundle.imports.sort(), [ '@scoped/bar', '@scoped/foo' ] );
-		});
+		}).then(getBundleImports)
+			.then(imports => assert.deepEqual(imports, ['@scoped/foo', '@scoped/bar']));
 	});
 
 	it( '"only" option works with a regex', () => {
@@ -611,9 +617,8 @@ describe( 'rollup-plugin-node-resolve', function () {
 					only: [ /^@scoped\/.*$/ ]
 				})
 			]
-		}).then(bundle => {
-			assert.deepEqual( bundle.imports.sort(), [ 'test' ] );
-		});
+		}).then(getBundleImports)
+			.then(imports => assert.deepEqual(imports, ['test']));
 	});
 
 	it( 'allows custom options', () => {
@@ -626,7 +631,7 @@ describe( 'rollup-plugin-node-resolve', function () {
 			}) ]
 		}).then( bundle => {
 			assert.equal(
-				bundle.modules[0].id,
+				bundle.cache.modules[1].id,
 				path.resolve( __dirname, 'samples/custom-resolve-options/js_modules/foo.js' )
 			);
 		});
@@ -638,8 +643,25 @@ describe( 'rollup-plugin-node-resolve', function () {
 			plugins: [ nodeResolve({
 				modulesOnly: true
 			}) ]
-		}).then( bundle => {
-			assert.deepEqual( bundle.imports, [ 'foo/deep' ] );
+		}).then(getBundleImports)
+			.then(imports => assert.deepEqual(imports, ['foo/deep']));
+	});
+
+	it( 'generates manual chunks', () => {
+		const chunkName = 'mychunk';
+		return rollup.rollup({
+			input: 'samples/manualchunks/main.js',
+			experimentalCodeSplitting: true,
+			manualChunks: {
+				[ chunkName ]: [ 'simple' ]
+			},
+			plugins: [ nodeResolve() ]
+		}).then( bundle =>
+			bundle.generate({
+				format: 'esm',
+				chunkFileNames: '[name]',
+			})).then( generated => {
+			assert.ok(getChunksFromGenerated(generated).find(({fileName}) => fileName === chunkName));
 		});
 	});
 
@@ -672,4 +694,12 @@ describe( 'rollup-plugin-node-resolve', function () {
 		});
 	});
 
+	it('resolves dynamic imports', () => {
+		return rollup.rollup({
+			input: 'samples/dynamic/main.js',
+			inlineDynamicImports: true,
+			plugins: [ nodeResolve() ]
+		}).then(executeBundle)
+			.then(({exports}) => exports.then(result => assert.equal(result.default, 42)));
+	});
 });