+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # rollup-plugin-node-resolve changelog
 
+## 4.0.0 (2018-12-09)
+
+This release will support rollup@1.0
+
+### Features
+* Resolve modules used to define manual chunks ([#185](https://github.com/rollup/rollup-plugin-node-resolve/pull/185) by @mcshaman)
+* Update dependencies and plugin hook usage ([#187](https://github.com/rollup/rollup-plugin-node-resolve/pull/187) by @lukastaegert)
+
 ## 3.4.0 (2018-09-04)
 
 This release now supports `.mjs` files by default